+++ lib/i18next.dart
@@ -1,5 +1,7 @@
 library i18next;
 
+export 'src/data_sources/asset_bundle_data_source.dart';
+export 'src/data_sources/localization_data_source.dart';
 export 'src/i18next.dart';
 export 'src/i18next_localization_delegate.dart';
 export 'src/options.dart';