+++ lib/src/data_sources/localization_data_source.dart
@@ -1,0 +1,5 @@
+import 'dart:ui';
+
+abstract class LocalizationDataSource {
+  Future<Map<String, Object>> load(Locale locale);
+}