+++ lib/src/i18next.dart
@@ -9,7 +9,7 @@ import 'resource_store.dart';
 import 'translator.dart';
 
 /// It translates the i18next localized format in your localization objects
-/// (provided by [dataSource]) to the final translation.
+/// (provided by [resourceStore]) to the final translation.
 ///
 /// Usually the most common usage:
 ///
@@ -40,9 +40,7 @@ class I18Next {
         pluralResolver = PluralResolver(),
         options = I18NextOptions.base.apply(options);
 
-  /// The current [Locale] for this instance.
-  ///
-  /// It is used as the default locale for pluralization rules and [dataSource].
+  /// The current and default [Locale] for this instance.
   final Locale locale;
 
   /// The resources store that contains all the necessary values mapped by